refactor(header): move header component into globalheader recipe
